pyproject.toml
[project]
name = "airflow-dag-project"
version = "0.1.0"
description = "Production-oriented Airflow DAG workspace with local debug, Docker Compose, and Kubernetes Spark examples."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "apache-airflow[cncf.kubernetes]==2.10.5",
  "PyYAML>=6.0.2,<7.0.0",
  "structlog>=25.4.0,<26.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5,<9.0.0",
  "pytest-cov>=6.1.1,<7.0.0",
  "requests>=2.32.3,<3.0.0",
  "ruff>=0.11.7,<0.12.0",
  "testcontainers[postgres]>=4.10.0,<5.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_dag_project"]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "-ra"
markers = [
  "integration: exercises Docker/Testcontainers-backed flows",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

